+++ changelog.md
@@ -1,6 +1,8 @@
 # Version 2023.10.9 (2023-10-12)
 
 - Fix register_update_callback for update
+- Add filter options to device.get_entity\*
+- Send relevant entities instead of devices in callback
 
 # Version 2023.10.8 (2023-10-11)